# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skool.loc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/"
        }
    ],
    "paths": {
...
        "/send-message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "New Chat",
                "description": "Chat yozish",
                "operationId": "c481e3979d92fccef07c2f192dac4d7e",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Chat",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "text",
                            "example": "Xabar yozish"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "number",
                            "example": "1"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "text",
                            "example": " token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Chat"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/get-chat-messages/{user_id}/{paginate}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "GetChatMessage",
                "description": "Get Course By Group ID",
                "operationId": "8f54e707add09c73f0cdf7f34a45043d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id of Course",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "path",
                        "description": "id of Course",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Chat"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/get-chat-users": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "GetChatUser",
                "description": "Get Course By Group ID",
                "operationId": "e564d16186695380dfcf826852b00694",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Chat"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Chat": {
                "required": [],
                "properties": {
                    "user_id": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": true
                    },
                    "sender_id": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "admin": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}